chore(code-garden): split BookmarksTab into container + sub-components [W30-2B]
Non-functional change from repo audit 2026-W30 (Finding 2-B).
Behavior preserved; all 175 mission-control tests pass unchanged.

Container (BookmarksTab.jsx): 811 → 167 lines. AC: `Container ≤ 200
lines` met.

Findings audited as `garden-safe` (pure structural decomposition,
behavior-preserving). Sub-files:

- BookmarksSankeyChart.jsx — was SankeySection (~160 lines, owns
  d3-sankey layout + chart render + own useState/useEffect for width).
- BookmarksStatesOverTimeChart.jsx — was StateCountsChart (~210 lines,
  owns line chart render + crosshair tooltip state).
- BookmarksPendingApprovals.jsx — extracted the pending approvals Card
  (~25 lines; renders nothing when array empty).
- BookmarksFunnelList.jsx — extracted the funnel Card (~35 lines;
  owns funnelMax calculation).
- BookmarksCurationsCard.jsx + BookmarksCurationItemRow.jsx — extracted
  the curations Card + its row helper (~95 lines combined).
- BookmarksTopicsCard.jsx — extracted the by-topic Card (~28 lines;
  owns topicMax calculation).
- BookmarksTransitionsCard.jsx — extracted the recent transitions Card
  (~35 lines).
- BookmarksKpiRow.jsx — extracted the KPI row (~28 lines).
- BookmarksSankeyCard.jsx — extracted the Sankey wrapper Card with
  Expand/Collapse toggle (~32 lines).
- BookmarksStateCountsCard.jsx — extracted the state-counts chart
  wrapper Card (~15 lines).
- BookmarksToolbar.jsx — extracted the toolbar + Field/Select helpers
  (~55 lines).

All data-testid attributes preserved — BookmarksTab.test.jsx passes
unchanged.
